server/requirements.txt
fastapi==0.79.0
uvicorn==0.18.3
PyJWT==2.6.0
python-decouple==3.8
pydantic==1.9.1
pydantic[email]
pymongo==4.1.1
pymongo[srv]
passlib==1.7.4
pandas==1.5.3
requests==2.28.2
bcrypt==4.0.1
scikit-learn==1.2.2